requirements.txt
pydantic
openai
scipy
python-dotenv